♻️ refactor: extract Tooltip into modular components
Extract the monolithic Tooltip component into separate, focused modules:
- TooltipInGroup: handles grouped tooltip behavior
- TooltipStandalone: handles standalone tooltip behavior
- useMergedTooltipProps: hooks for merging tooltip props
- useTooltipFloating: floating UI positioning logic
- useTooltipReference: reference element management
- useTooltipTrigger: trigger behavior management

This refactoring improves code organization, maintainability, and reusability.
